requirements.txt
numpy==1.16.4
opencv-contrib-python==4.1.0.25
python-dotenv==0.10.3
pathlib2==2.3.4
pytest==5.1.3
pytest-cov==2.7.1
pylint==2.4.4
python-json-logger==0.1.11
joblib==0.14.1